+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/MiniAppDownloader.kt
@@ -194,8 +194,7 @@ internal class MiniAppDownloader(
         }
     }
 
-    fun getDownloadedMiniAppList(): List<MiniAppInfo> =
-        miniAppStatus.getDownloadedMiniAppList()
+    fun getDownloadedMiniAppList(): List<MiniAppInfo> = miniAppStatus.getDownloadedMiniAppList()
 
     @Suppress("SENSELESS_COMPARISON")
     @VisibleForTesting